Drop `ParameterTrait`
